Add support for new layout in compose
