+++ gradle.properties
@@ -13,7 +13,6 @@ org.gradle.jvmargs=-Xmx1536m
 
 org.gradle.parallel=true
 
-android.enableJetifier=true
 android.useAndroidX=true
 
 android.defaults.buildfeatures.buildconfig=false
