+++ library/src/main/java/com/transferwise/sequencelayout/SequenceAdapter.kt
@@ -1,13 +1,13 @@
 package com.transferwise.sequencelayout
 
 /**
- * Adapter to bind data to [com.transferwise.sequencelayout.SequenceStep]s for a [com.transferwise.sequencelayout.SequenceLayout].
+ * Adapter to bind data to [com.transferwise.sequencelayout.SequenceStepView]s for a [com.transferwise.sequencelayout.SequenceLayoutView].
  */
-public abstract class SequenceAdapter<T> where T : Any {
+abstract class SequenceAdapter<T> where T : Any {
 
     abstract fun getCount(): Int
 
     abstract fun getItem(position: Int): T
 
-    abstract fun bindView(sequenceStep: SequenceStep, item: T)
+    abstract fun bindView(sequenceStep: SequenceStepView, item: T)
 }
