+++ sample/src/main/res/layout/activity_main.xml
@@ -15,23 +15,23 @@
         android:padding="16dp"
         tools:ignore="HardcodedText">
 
-        <com.transferwise.sequencelayout.SequenceLayout
+        <com.transferwise.sequencelayout.SequenceLayoutView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginTop="16dp">
 
-            <com.transferwise.sequencelayout.SequenceStep
+            <com.transferwise.sequencelayout.SequenceStepView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 app:anchor="30 Nov"
                 app:title="First step"/>
 
-            <com.transferwise.sequencelayout.SequenceStep
+            <com.transferwise.sequencelayout.SequenceStepView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 app:title="Second step"/>
 
-            <com.transferwise.sequencelayout.SequenceStep
+            <com.transferwise.sequencelayout.SequenceStepView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:paddingBottom="8dp"
@@ -42,17 +42,17 @@
                 app:title="Third step"
                 app:titleTextAppearance="@style/TextAppearance.AppCompat.Title"/>
 
-            <com.transferwise.sequencelayout.SequenceStep
+            <com.transferwise.sequencelayout.SequenceStepView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 app:title="Fourth step"/>
 
-            <com.transferwise.sequencelayout.SequenceStep
+            <com.transferwise.sequencelayout.SequenceStepView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 app:anchor="2 Dec"
                 app:title="Fifth step"/>
-        </com.transferwise.sequencelayout.SequenceLayout>
+        </com.transferwise.sequencelayout.SequenceLayoutView>
 
     </LinearLayout>
 </ScrollView>
